+++ packages/middleware-sentry/src/index.ts
@@ -1,148 +1,1 @@
-import * as Sentry from "@sentry/node";
-import { InngestMiddleware } from "inngest";
-
-/**
- * Options used to configure the Sentry middleware.
- */
-export interface SentryMiddlewareOptions {
-  /**
-   * If `true`, the Sentry middleware will not automatically flush events after
-   * each function run. This can be useful if you want to control when events
-   * are sent to Sentry, or leave it to Sentry's default behavior.
-   *
-   * By default, automatic flushing is enabled to ensure that events are sent in
-   * serverless environments where the runtime may be terminated if the function
-   * has returned a value.
-   *
-   * @default false
-   */
-  disableAutomaticFlush?: boolean;
-}
-
-/**
- * Captures errors and performance data from Inngest functions and sends them to
- * Sentry.
- *
- * This imports Sentry directly and relies on it already being initialized using
- * `Sentry.init()`. For more information on how to configure Sentry, see the
- * [Sentry documentation](https://docs.sentry.io/platforms/node/).
- */
-export const sentryMiddleware = (
-  /**
-   * Options used to configure the Sentry middleware.
-   */
-  opts?: SentryMiddlewareOptions
-) => {
-  return new InngestMiddleware({
-    name: "@inngest/middleware-sentry",
-    init({ client }) {
-      return {
-        onFunctionRun({ ctx, fn, steps }) {
-          return Sentry.withScope((scope) => {
-            scope.setTags({
-              "inngest.client.id": client.id,
-              "inngest.function.id": fn.id(client.id),
-              "inngest.function.name": fn.name,
-              "inngest.event": ctx.event.name,
-              "inngest.run.id": ctx.runId,
-            });
-
-            let memoSpan: Sentry.Span;
-            let execSpan: Sentry.Span;
-
-            return Sentry.startSpanManual(
-              {
-                name: "Inngest Function Run",
-                op: "run",
-                attributes: { "inngest.event": JSON.stringify(ctx.event) },
-              },
-              (reqSpan) => {
-                return {
-                  transformInput() {
-                    return {
-                      ctx: {
-                        /**
-                         * The Sentry client passed to
-                         * `@inngest/middleware-sentry`.
-                         */
-                        sentry: Sentry,
-                      },
-                    };
-                  },
-                  beforeMemoization() {
-                    Sentry.withActiveSpan(reqSpan, () => {
-                      Sentry.startSpanManual(
-                        {
-                          name: "Memoization",
-                          op: "memoization",
-                          attributes: {
-                            "inngest.memoization.count": steps.length,
-                          },
-                        },
-                        (_memoSpan) => {
-                          memoSpan = _memoSpan;
-                        }
-                      );
-                    });
-                  },
-                  afterMemoization() {
-                    memoSpan?.end();
-                  },
-                  beforeExecution() {
-                    Sentry.withActiveSpan(reqSpan, () => {
-                      Sentry.startSpanManual(
-                        {
-                          name: "Execution",
-                          op: "execution",
-                        },
-                        (_execSpan) => {
-                          execSpan = _execSpan;
-                        }
-                      );
-                    });
-                  },
-                  afterExecution() {
-                    execSpan?.end();
-                  },
-                  transformOutput({ result, step }) {
-                    // Set step metadata
-                    if (step) {
-                      Sentry.withActiveSpan(reqSpan, (scope) => {
-                        scope.setTags({
-                          "inngest.step.name": step.displayName,
-                          "inngest.step.op": step.op,
-                        });
-                      });
-                    }
-
-                    // Capture step output and log errors
-                    if (result.error) {
-                      reqSpan.setStatus({
-                        code: 2,
-                      });
-
-                      Sentry.withActiveSpan(reqSpan, (scope) => {
-                        scope.captureException(result.error);
-                      });
-                    } else {
-                      reqSpan.setStatus({
-                        code: 1,
-                      });
-                    }
-                  },
-                  async beforeResponse() {
-                    reqSpan.end();
-
-                    if (!opts?.disableAutomaticFlush) {
-                      await Sentry.flush();
-                    }
-                  },
-                };
-              }
-            );
-          });
-        },
-      };
-    },
-  });
-};
+export * from "./middleware";
